Add files via upload
User Analytics in the Telecommunication
Industry
You are working for a wealthy investor that specializes in purchasing assets that are
undervalued. This investor’s due diligence on all purchases includes a detailed analysis of
the data that underlies the business, to try to understand the fundamentals of the business
and especially to identify opportunities to drive profitability by changing the focus of which
products or services are being offered.
Your last role with this investor saw you do a rich analysis of a delivery company and you
helped to identify that delivery to university students was the most profitable route to follow,
and your analysis helped the investor purchase this delivery company and ramp up profits by
25% within 6 months through focussing on the most profitable aspect of the business. This
was driven by university students always being hungry, awake at all hours, willing to
purchase from a limited food menu and tending to live within a small geographical area.
The investor is interested in purchasing TellCo, an existing mobile service provider in the
Republic of Pefkakia. TellCo’s current owners have been willing to share their financial
information but have never employed anyone to look at their system-generated data.
Your employer wants you to provide a report to analyze opportunities for growth and make a
recommendation on whether TellCo is worth buying or selling. You will do this by analyzing a
telecommunication dataset that contains useful information about the customers & their
activities on the network. You will deliver insights you managed to extract to your employer
through an easy-to-use web-based dashboard and a written report.